🚚refactor: auth 패키지 수정에 따른 코드 수정
