+++ src/main/java/com/project/openmarket/domain/auth/enums/SessionConst.java
@@ -1,4 +1,4 @@
-package com.project.openmarket.global.auth.enums;
+package com.project.openmarket.domain.auth.enums;
 
 public class SessionConst {
 	public static final String SESSION_KEY = "LoginUser";
