+++ src/test/java/com/project/openmarket/global/interceptor/ConsumerInterceptorTests.java
@@ -17,8 +17,8 @@ import org.springframework.mock.web.MockHttpSession;
 import com.project.openmarket.domain.user.dto.request.ConsumerCreateReqestDto;
 import com.project.openmarket.domain.user.entity.Consumer;
 import com.project.openmarket.domain.user.repository.ConsumerRepository;
-import com.project.openmarket.global.auth.ConsumerThreadLocal;
-import com.project.openmarket.global.auth.enums.SessionConst;
+import com.project.openmarket.domain.auth.ConsumerThreadLocal;
+import com.project.openmarket.domain.auth.enums.SessionConst;
 
 class ConsumerInterceptorTests extends InterceptorTestMock {
 	@InjectMocks
